Renamed client/client.ts to hooks, because it has only hooks and types Client/client made no sense. There is already client/index and oidc-react and client.ts just makes it harder to find things. Removed also unnecessary eslint-disable.
